renamed deps.py to ninja.py and removed old ninja.py
