+++ ninja.py
@@ -1,59 +1,244 @@
-import sys
+import functools
 
-HEADER = """# This file is automatically generated
-"""
+def Dep(inputs, outputs, rule, params=None):
+    return {
+        'inputs': inputs,
+        'outputs': outputs,
+        'rule': rule,
+        'params': params,
+    }
 
-wordnet_deps = ['raw/wordnet/'+path for path in
-    ['wordnet-synset.ttl','wordnet-glossary.ttl',
-    'full/wordnet-wordsense-synset-relations.ttl',
-    'wordnet-attribute.ttl', 'wordnet-causes.ttl',
-    'wordnet-classifiedby.ttl', 'wordnet-entailment.ttl',
-    'wordnet-hyponym.ttl', 'wordnet-instances.ttl',
-    'wordnet-membermeronym.ttl', 'wordnet-partmeronym.ttl',
-    'wordnet-sameverbgroupas.ttl', 'wordnet-similarity.ttl',
-    'wordnet-substancemeronym.ttl', 'full/wordnet-antonym.ttl',
-    'full/wordnet-derivationallyrelated.ttl',
-    'full/wordnet-participleof.ttl',
-    'full/wordnet-pertainsto.ttl',
-    'full/wordnet-seealso.ttl']]
+prefix = 'data/'
+version = '5.3'
 
-globalmind_deps = ['raw/globalmind/GMFrame.yaml',
-     'raw/globalmind/GMTranslation.yaml',
-     'raw/globalmind/GMAssertion.yaml',
-     'raw/globalmind/GMUser.yaml',]
+wiktionary_langs = ['en', 'de', 'ja']
+wiktionary_slices = 20
+collate_count = 8
 
-from_tar_file = [
- 'raw/dbpedia/mappingbased_properties_en.nt',
- 'raw/dbpedia/interlanguage_links_en.nt',
- 'raw/dbpedia/instance_types_en.nt',
-] + \
-    ['raw/jmdict/JMdict.xml',] + \
-    ['raw/wiktionary/enwiktionary.xml',
-    'raw/wiktionary/dewiktionary.xml',
-    'raw/wiktionary/jawiktionary.xml',] + \
-    ['raw/conceptnet4%s/conceptnet4%s_flat_%s.jsons'%(ext, ext, i)
-    for i in range(10) for ext in ['', '_nadya']] + \
-    ['raw/conceptnet_zh/conceptnet_zh_part%s.txt'%i
-    for i in range(13)] + \
-    wordnet_deps + \
-    ['raw/verbosity/verbosity.txt',] + \
-    ['raw/umbel/umbel.nt',] + \
-    globalmind_deps
+in_tar = {
+    'wordnet':
+        ['wordnet-synset.ttl','wordnet-glossary.ttl',
+        'full/wordnet-wordsense-synset-relations.ttl',
+        'wordnet-attribute.ttl', 'wordnet-causes.ttl',
+        'wordnet-classifiedby.ttl', 'wordnet-entailment.ttl',
+        'wordnet-hyponym.ttl', 'wordnet-instances.ttl',
+        'wordnet-membermeronym.ttl', 'wordnet-partmeronym.ttl',
+        'wordnet-sameverbgroupas.ttl', 'wordnet-similarity.ttl',
+        'wordnet-substancemeronym.ttl', 'full/wordnet-antonym.ttl',
+        'full/wordnet-derivationallyrelated.ttl',
+        'full/wordnet-participleof.ttl',
+        'full/wordnet-pertainsto.ttl',
+        'full/wordnet-seealso.ttl'],
+
+    'globalmind':
+        ['GMFrame.yaml',
+         'GMTranslation.yaml',
+         'GMAssertion.yaml',
+         'GMUser.yaml',],
+
+    'dbpedia':
+        ['mappingbased_properties_en.nt',
+        'interlanguage_links_en.nt',
+        'instance_types_en.nt',],
+
+    'jmdict':
+        ['JMdict.xml'],
+
+    'wiktionary':
+        ['enwiktionary.xml',
+        'dewiktionary.xml',
+        'jawiktionary.xml',],
+
+    'conceptnet4': ['conceptnet4_flat_%s.jsons'%i
+        for i in range(10)],
+
+    'conceptnet4_nadya': ['conceptnet4_nadya_flat_%s.jsons'%i
+        for i in range(10)],
+
+    'conceptnet_zh': ['rconceptnet_zh_part%s.txt'%i
+        for i in range(13)],
+
+    'verbosity': ['raw/verbosity/verbosity.txt',],
+    'umbel': ['raw/umbel/umbel.nt',],
+
+}
+
+in_tar = {k: [prefix+'raw/%s/%s'%(k, file) for file in v]
+            for k,v in in_tar.items()}
+
+def add_all_deps(deps):
+    download(deps)
+    untar(deps)
+    parse_sw(deps) #wordnet, umbel
+    parse_standard(deps) #globalmind, jmdict, verbosity
+    parse_conceptnet4(deps) #conceptnet4 conceptnet4_nadya conceptnet_zh
+    extract_wiktionary(deps)
+    parse_wiktionary(deps)
+    msgpack_to_csv(deps)
+    collate(deps)
+    count_and_rank(deps)
+    combine_assertions(deps)
+    build_db(deps)
+
+def download(deps):
+    output = prefix+'conceptnet5_raw_data_%s.tar.bz2'%version
+    url = 'http://conceptnet5.media.mit.edu/downloads/v%s/'%version+output
+    deps['download_tar'] = Dep(
+        [],
+        prefix+'conceptnet5_raw_data_%s.tar.bz2'%version,
+        'download',
+        {'prefix': prefix, 'url': url}
+    )
+
+def untar(deps):
+    outputs = []
+    for files in in_tar.values():
+        outputs += files
+    input = deps['download_tar']['outputs']
+    deps['untar'] = Dep([input], outputs, 'extract_tar', {'prefix': prefix})
+
+def parse_sw(deps):
+    for type in ['wordnet', 'umbel']:
+        deps['parse %s'%type] = Dep(
+            in_tar[type],
+            edge_output_str(type),
+            'parse_sw',
+            {
+                'parser': type,
+                'prefix': prefix+'sw_map',
+                'dir': prefix+'raw/'+type
+            })
+
+def parse_standard(deps):
+    for type in ['globalmind', 'jmdict', 'verbosity']:
+        deps['parse %s'%type] = Dep(
+            in_tar[type],
+            edge_output_str(type),
+            'parse',
+            {'parser': type})
+
+def parse_conceptnet4(deps):
+    for type in ['conceptnet4', 'conceptnet4_nadya', 'conceptnet_zh']:
+        for input in in_tar[type]:
+            output = input.replace('jsons', 'msgpack')\
+                        .replace('txt', 'msgpack')\
+                        .replace('raw', 'edges')
+            parser = type if not type.endswith('zh') else 'ptt_petgame'
+            deps['parse %s'%type] = Dep(
+                [input],
+                [output],
+                'parse',
+                {'parser': parser})
 
 
-def add_dep(lines, rule, input, output, extra=None, params=None):
-    if isinstance(output, list):
-        output = ' '.join(output)
-    if isinstance(input, list):
-        input = ' '.join(input)
+def extract_wiktionary(deps):
+    for lang in wiktionary_langs:
+        input = prefix+'raw/wiktionary/%swiktionary.xml'%lang
+
+        outputs = [prefix+'extracted/wiktionary/%s/wiktionary_%02d.msgpack'%(lang, i)
+                    for i in range(wiktionary_slices)]
+
+        deps['extract %s wiktionary'%lang] = Dep(
+            [input],
+            outputs,
+            'extract_wiktionary',
+            {'lang': lang})
+
+def parse_wiktionary(deps):
+    for lang in wiktionary_langs:
+        for input in deps['extract %s wiktionary'%lang]['outputs']:
+            output = input.replace('extracted', 'edges')
+            index = input[-10:-8]
+
+            deps['parse %s wiktionary %s'%(lang, index)] = {
+                'inputs': [input],
+                'outputs': [input.replace('extracted', 'edges')],
+                'rule': 'parse_wiktionary',
+                'params': {'lang': lang}
+            }
+
+def msgpack_to_csv(deps):
+    new_deps = {}
+    for k,v in deps.items():
+        if not k.startswith('parse'): continue
+
+        for input in v['outputs']:
+            new_deps['msgpack to csv %s'%input] = Dep(
+                [input],
+                [input.replace('msgpack', 'csv')],
+                'msgpack_to_csv')
+
+    deps.update(new_deps)
+
+def collate(deps):
+    inputs = []
+    for k,v in deps.items():
+        if k.startswith('msgpack to csv'):
+            inputs += v['outputs']
+
+    deps['collate'] = Dep(
+        inputs,
+        [prefix+'edges/split/edges_%02d.csv'%i
+            for i in range(collate_count)],
+        'collate',
+        {'count': collate_count}
+    )
+
+def count_and_rank(deps):
+    for input in deps['collate']['outputs']:
+        deps['count and rank %s'%input] = Dep(
+            [input],
+            [input.replace('split', 'sorted')],
+            'count_and_rank')
+
+def combine_assertions(deps):
+    new_deps = {}
+    for k,v in deps.items():
+        if not k.startswith('count and rank'): continue
+        input = v['outputs'][0] #will only be a single element list
+
+        new_deps['combine assertions %s'%input] = Dep(
+            [input],
+            [input.replace('edges', 'assertions').replace('csv', 'msgpack')],
+            'combine_assertions')
+
+    deps.update(new_deps)
+
+def build_db(deps):
+    inputs = []
+    for k,v in deps.items():
+        if not k.startswith('combine assertions'): continue
+        inputs += v['outputs']
+    output = prefix+'db/assertions.db'
+
+    deps['build db'] = Dep(
+        inputs,
+        [output],
+        'build_db'
+    )
+
+def edge_output_str(type):
+    return [prefix+'edges/%s/%s.msgpack'%(type, type)]
+
+def to_ninja(rules, deps):
+    lines = [rules]
+    for dep in deps.values():
+        add_dep(lines, **dep)
+    return "\n".join(lines)
+
+def add_dep(lines, rule, inputs, outputs, extra=None, params=None):
+    if isinstance(outputs, list):
+        outputs = ' '.join(outputs)
+    if isinstance(inputs, list):
+        inputs = ' '.join(inputs)
     if extra:
         if isinstance(extra, list):
             extra = ' '.join(extra)
         extrastr = ' | ' + extra
     else:
         extrastr = ''
-    build_rule = "build {output}: {rule} {input}{extra}".format(
-        output=output, rule=rule, input=input, extra=extrastr
+    build_rule = "build {outputs}: {rule} {inputs}{extra}".format(
+        outputs=outputs, rule=rule, inputs=inputs, extra=extrastr
     )
     lines.append(build_rule)
     if params:
@@ -61,131 +246,12 @@ def add_dep(lines, rule, input, output, extra=None, params=None):
             lines.append("  {key} = {val}".format(key=key, val=val))
     lines.append("")
 
-def write_ninja_deps(rules_filename='rules.ninja', out=sys.stdout, version='5.3',
-        prefix='data/'):
-
-    print(HEADER, file=out)
-
-    with open(rules_filename, encoding='utf-8') as rulesfile:
-        print(rulesfile.read(), file=out)
-
-    lines = []
-
-    lines.extend(download_deps(version, prefix))
-    lines.extend(unpack_deps(version, prefix))
-    lines.extend(parse_jmdict(prefix))
-    lines.extend(parse_wiktionary(prefix))
-    lines.extend(parse_conceptnet4(prefix))
-    lines.extend(parse_ptt_petgame(prefix))
-    lines.extend(parse_wordnet(prefix))
-    lines.extend(parse_umbel(prefix))
-    lines.extend(parse_verbosity(prefix))
-    lines.extend(parse_globalmind(prefix))
-
-    print('\n'.join(lines), file=out)
-
-def download_deps(version, prefix):
-    lines = []
-    output = prefix+'conceptnet5_raw_data_%s.tar.bz2'%version
-    url = 'http://conceptnet5.media.mit.edu/downloads/v%s/'%version+output
-    add_dep(lines, 'download', '', output, params={'prefix': prefix, 'url': url})
-    return lines
-
-#TODO Add other outputs
-def unpack_deps(version, prefix):
-    lines = []
-    input = prefix+'conceptnet5_raw_data_%s.tar.bz2'%version
-
-    output = from_tar_file
-    output = [prefix+out for out in output]
-
-    add_dep(lines, 'extract_tar', input, output, params={'prefix': prefix})
-    return lines
-
-def parse_jmdict(prefix):
-    lines = []
-    input = prefix+'raw/jmdict/JMdict.xml'
-    output = prefix+'/edges/jmdict/jmdict.msgpack'
-    add_dep(lines, 'parse', input, output, params={'filetype': 'jmdict'})
-    return lines
-
-def parse_wiktionary(prefix, langs=('en', 'de', 'ja'), slices=20):
-    lines = []
-    for lang in langs:
-        input = prefix+'raw/wiktionary/%swiktionary.xml'%lang
-        output_prefix = prefix+'extracted/wiktionary/%s'%lang
-        output = [output_prefix+'wiktionary_%02d.msgpack'%i for i in range(slices)]
-        add_dep(lines, 'extract_wiktionary', input, output, params={'lang':lang})
-
-        for i, new_input in enumerate(output):
-            new_output = prefix+'edges/wiktionary/%s/wiktionary_%02d.msgpack'%(lang, i)
-            add_dep(lines, 'parse_wiktionary', new_input, new_output, params={'lang':lang})
-
-    return lines
-
-#parsers both conceptnet4 and conceptnet4 nadya
-def parse_conceptnet4(prefix):
-    lines = []
-    for style in ['conceptnet4', 'conceptnet4_nadya']:
-        for i in range(10):
-            input = prefix+'raw/%s/%s_flat_%s.jsons'%(style, style, i)
-            output = prefix+'edges/%s%s_flat_%s.msgpack'%(style, style, i)
-            add_dep(lines, 'parse', input, output, params={'filetype':'conceptnet4'})
-
-    return lines
-
-
-def parse_ptt_petgame(prefix):
-    lines = []
-    for i in range(13):
-        input = prefix+'raw/conceptnet_zh/conceptnet_zh_part%s.txt'%i
-        output = prefix+'edges/conceptnet_zh/conceptnet_zh_part%s.csv'%i
-        add_dep(lines, 'parse', input, output, params={'filetype': 'ptt_petgame'})
-
-    return lines
-
-def parse_wordnet(prefix):
-    lines = []
-    input = [prefix+dep for dep in wordnet_deps]
-    output = ['edges/wordnet/wordnet.msgpack', 'sw_map/wordnet.nt']
-    output = [prefix+out for out in output]
-    params = {'filetype': 'wordnet', 'prefix': prefix+'sw_map', 'dir': prefix+'raw/wordnet'}
-    add_dep(lines, 'parse_sw', input, output, params=params)
-
-    return lines
-
-def parse_umbel(prefix):
-    lines = []
-    input = prefix+'raw/umbel/umbel.nt'
-    output = ['edges/umbel/umbel.msgpack', 'sw_map/umbel.nt']
-    output = [prefix+out for out in output]
-    params = {'filetype': 'umbel', 'prefix': prefix+'sw_map', 'dir':prefix+'raw/umbel'}
-    add_dep(lines, 'parse_sw', input, output, params=params)
-
-    return lines
-
-def parse_verbosity(prefix):
-    lines = []
-    input = prefix+'raw/verbosity/verbosity.txt'
-    output = prefix+'edges/verbosity/verbosity.msgpack'
-    add_dep(lines, 'parse', input, output, params={'filetype': 'verbosity'})
-
-    return lines
-
-def parse_globalmind(prefix):
-    lines = []
-    input = [prefix+dep for dep in globalmind_deps]
-    output = prefix+'edges/globalmind/globalmind.msgpack'
-    add_dep(lines, 'parse', input, output, params={'filetype': 'verbosity'})
-
-    return lines
-
-
-def convert_to_csv(prefix):
-    pass
-
 def main():
-    write_ninja_deps(out=open('build.ninja', mode='w'))
+    deps = {}
+    add_all_deps(deps)
+    ninja = to_ninja(open('rules.ninja').read(), deps)
+    print(ninja, file=open('build.ninja', mode='w'))
+
 
 if __name__ == '__main__':
     main()
